WIP: core: Add WatchdogDeviceNum setting
This number counts from zero, specifying the watchdog device to use. By
default the value is zero, which preserves the existing behaviour.

If a system requires a watchdog other than /dev/watchdog0 to be used to
reboot the sytem, this setting schould be changed to the relevant
watchdog number.

Compile tested. Not complete.